requirements.txt
openai-whisper
torch
requests
tqdm
